+++ changelog.md
@@ -1,6 +1,13 @@
 ====================================================================
 CHANGELOG
 ====================================================================
+# 2.0.7
+* NEW: Add hero button target.
+* FIXED: Duplicate title on single post.
+* FIXED: Duplicate title on product archive.
+* FIXED: Contact form issue.
+
+
 ## 2.0.6
 * FIXED: JS error issue.